Move the views to their own module
