+++ tests/test_flask_oidc.py
@@ -67,49 +67,6 @@ def test_signin(test_app, client, mocked_responses, dummy_token):
     assert resp.get_data(as_text=True) == "dummy_refresh_token"
 
 
-def test_authorize_error(client):
-    resp = client.get(
-        "http://localhost/authorize?error=dummy_error&error_description=Dummy+Error"
-    )
-    assert resp.status_code == 401
-    assert "<p>dummy_error: Dummy Error</p>" in resp.get_data(as_text=True)
-
-
-def test_authorize_no_return_url(client, mocked_responses, dummy_token):
-    mocked_responses.post("https://test/openidc/Token", json=dummy_token)
-    mocked_responses.get("https://test/openidc/UserInfo", json={"nickname": "dummy"})
-    with client.session_transaction() as session:
-        session["_state_oidc_dummy_state"] = {"data": {}}
-    resp = client.get("/authorize?state=dummy_state&code=dummy_code")
-    assert resp.status_code == 302
-    assert resp.location == "http://localhost/"
-
-
-def test_authorize_no_user_info(test_app, client, mocked_responses, dummy_token):
-    test_app.config["OIDC_USER_INFO_ENABLED"] = False
-    mocked_responses.post("https://test/openidc/Token", json=dummy_token)
-    with client.session_transaction() as session:
-        session["_state_oidc_dummy_state"] = {"data": {}}
-    resp = client.get("/authorize?state=dummy_state&code=dummy_code")
-    assert resp.status_code == 302
-    assert "oidc_auth_token" in flask.session
-    assert "oidc_auth_profile" not in flask.session
-
-
-def test_logout(client, dummy_token):
-    with client.session_transaction() as session:
-        session["oidc_auth_token"] = dummy_token
-        session["oidc_auth_profile"] = {"nickname": "dummy"}
-    resp = client.get("/logout")
-    assert resp.status_code == 302
-    assert resp.location == "http://localhost/"
-    assert "oidc_auth_token" not in flask.session
-    assert "oidc_auth_profile" not in flask.session
-    flashes = flask.get_flashed_messages()
-    assert len(flashes) == 1
-    assert flashes[0] == "You were successfully logged out."
-
-
 def test_ext_logout(test_app, client, dummy_token):
     with test_app.test_request_context(path="/somewhere"):
         flask.session["oidc_auth_token"] = dummy_token
@@ -117,16 +74,6 @@ def test_ext_logout(test_app, client, dummy_token):
         with pytest.warns():
             resp = test_app.oidc_ext.logout(return_to="/somewhere_else")
     assert resp.location == "/logout?next=/somewhere_else"
-
-
-def test_logout_expired(client, dummy_token):
-    with client.session_transaction() as session:
-        session["oidc_auth_token"] = dummy_token
-        session["oidc_auth_profile"] = {"nickname": "dummy"}
-    client.get("/logout?reason=expired")
-    flashes = flask.get_flashed_messages()
-    assert len(flashes) == 1
-    assert flashes[0] == "Your session expired, please reconnect."
 
 
 def test_expired_token(client, dummy_token):
@@ -149,13 +96,6 @@ def test_bad_token(client):
     assert "oidc_auth_token" not in flask.session
     assert "oidc_auth_profile" not in flask.session
     assert "TypeError: string indices must be integers" in resp.get_data(as_text=True)
-
-
-def test_oidc_callback_route(test_app, client, dummy_token):
-    with pytest.warns(DeprecationWarning):
-        resp = client.get("/oidc_callback?state=dummy-state&code=dummy-code")
-    assert resp.status_code == 302
-    assert resp.location == "/authorize?state=dummy-state&code=dummy-code"
 
 
 def test_user_getinfo(test_app, client, dummy_token):
